+++ index.html
@@ -34,6 +34,6 @@
         <img src="img/рубашка.png" width="120px" alt="back of a card">
         <img src="img/рубашка.png" width="120px" alt="back of a card">
     </div> -->
-    <script type="module" src="index.js"></script>
+    <script type="module" src="src/index.js"></script>
 </body>
 </html>
