+++ src/components.js
@@ -1,0 +1,55 @@
+export const header = `<div class="header">
+        <div class="header__timer">
+            <div class="header__timer__time__unit">
+                <p class="header__timer__time__unit__text">min</p>
+                <p class="header__timer__time__unit__text">sek</p>
+            </div>
+            <div class="header__timer__time">
+                <p>00.00</p>
+            </div>
+        </div>
+        <button class="header__button__retry">Начать заново</button>
+        </div>`;
+
+export let imagesCards = [
+    `<img src="img/6 бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/6 крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/6 черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/6 пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/7 бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/7 крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/7 черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/7 пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/8 бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/8 крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/8 черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/8 пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/9 бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/9 крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/9 черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/9 пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/10 бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/10 крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/10 черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/10 пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/валет бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/валет крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/валет черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/валет пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/дама бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/дама крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/дама черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/дама пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/король бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/король крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/король черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/король пики.png" width="120px" alt="back of a card">`,
+    `<img src="img/туз бубны.png" width="120px" alt="back of a card">`,
+    `<img src="img/туз крести.png" width="120px" alt="back of a card">`,
+    `<img src="img/туз черви.png" width="120px" alt="back of a card">`,
+    `<img src="img/туз пики.png" width="120px" alt="back of a card">`,
+];
+
+export const backOfCard = [
+    `<img src="img/рубашка.png" width="120px" alt="back of a card">`,
+];
