+++ src/index.js
@@ -1,20 +1,17 @@
-import { getLevelDifficulty } from "./renderLevels.js";
-import { getNoAnswer } from "./renderLevels.js";
+import { getLevelDifficulty } from './renderLevels.js';
+import { getNoAnswer } from './renderLevels.js';
 
 export let levelDifficulty;
 
 const renderLevelApp = (app, html) => {
-
-    return app.innerHTML = html;
-
-}
+    return (app.innerHTML = html);
+};
 
 const renderApp = () => {
+    const appElement = document.querySelector('body');
 
-    const appElement = document.querySelector("body");
+    appElement.classList.add('start');
 
-    appElement.classList.add('start')
-    
     let appHtml = `<div class="container">
             <div class="container__heading">
                 <p class="container__heading__text">Выбери сложность</p>
@@ -27,45 +24,40 @@ const renderApp = () => {
             <div class="container__start">
                 <button class="container__start__button">Старт</button>
             </div>
-        </div>`
-    
+        </div>`;
+
     appElement.innerHTML = appHtml;
 
-    const buttonLevelElements = document.querySelectorAll(".container__level__button");
-    const buttonStartElement = document.querySelector(".container__start__button");
+    const buttonLevelElements = document.querySelectorAll(
+        '.container__level__button',
+    );
+    const buttonStartElement = document.querySelector(
+        '.container__start__button',
+    );
 
     for (let i = 0; i < buttonLevelElements.length; i++) {
-
         buttonLevelElements[i].addEventListener('click', () => {
-
-            levelDifficulty = buttonLevelElements[i].dataset.level
-
+            levelDifficulty = buttonLevelElements[i].dataset.level;
         });
-        
-    };
+    }
 
     buttonStartElement.addEventListener('click', () => {
-
         appElement.classList.remove('start');
 
         // getLevelDifficulty(appHtml);
 
         if (!levelDifficulty) {
-            
             alert('Пожалуйста, выберите уровень');
 
             appElement.classList.add('start');
-
         } else {
-
             renderLevelApp(appElement, getLevelDifficulty());
-            
-            setTimeout(() => {return renderLevelApp(appElement, getNoAnswer())}, 5000);
-            
+
+            setTimeout(() => {
+                return renderLevelApp(appElement, getNoAnswer());
+            }, 5000);
         }
-
     });
-
 };
 
 renderApp();
